add draft and pubish item

+++ posts/views.py
@@ -1,34 +1,43 @@
+from urllib import quote_plus
+
+from django.utils import timezone
 from django.contrib import messages
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect, Http404
 from django.shortcuts import render, get_object_or_404, redirect
 
 from .forms import PostForm
 from .models import Post
 
 def post_create(request):
-	form = PostForm(request.POST or None, request.FILES or None)
-	if form.is_valid():
-		instance = form.save(commit=False)
-		instance.save()
-		# message success
-		messages.success(request, "Successfully Created")
-		return HttpResponseRedirect(instance.get_absolute_url())
-	context = {
-		"form": form,
-	}
-	return render(request, "post_form.html", context)
+    # Permissions user
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+    form = PostForm(request.POST or None, request.FILES or None)
+    if form.is_valid():
+        instance = form.save(commit=False)
+        instance.user = request.user
+        instance.save()
+        # message success
+        messages.success(request, "Successfully Created")
+        return HttpResponseRedirect(instance.get_absolute_url())
+    context = {
+    "form": form,
+    }
+    return render(request, "post_form.html", context)
 
 def post_detail(request, slug=None):
-	instance = get_object_or_404(Post, slug=slug)
-	context = {
-		"title": instance.title,
-		"instance": instance,
-	}
-	return render(request, "post_detail.html", context)
+    instance = get_object_or_404(Post, slug=slug)
+    share_string = quote_plus(instance.content)
+    context = {
+    	"title": instance.title,
+    	"instance": instance,
+        "share_string": share_string,
+    }
+    return render(request, "post_detail.html", context)
 
 def post_list(request):
-	queryset_list = Post.objects.all() #.order_by("-timestamp")
+	queryset_list = Post.objects.filter(draft=False).filter(publish__lte=timezone.now()) #.order_by("-timestamp")
 	paginator = Paginator(queryset_list, 10) # Show 25 contacts per page
 	page_request_var = "page"
 	page = request.GET.get(page_request_var)
@@ -50,28 +59,30 @@ def post_list(request):
 	return render(request, "post_list.html", context)
 
 
-
-
-
 def post_update(request, slug=None):
-	instance = get_object_or_404(Post, slug=slug)
-	form = PostForm(request.POST or None, request.FILES or None, instance=instance)
-	if form.is_valid():
-		instance = form.save(commit=False)
-		instance.save()
-		messages.success(request, "<a href='#'>Item</a> Saved", extra_tags='html_safe')
-		return HttpResponseRedirect(instance.get_absolute_url())
-
-	context = {
-		"title": instance.title,
-		"instance": instance,
-		"form":form,
-	}
-	return render(request, "post_form.html", context)
+    # Permissions user
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+    instance = get_object_or_404(Post, slug=slug)
+    form = PostForm(request.POST or None, request.FILES or None, instance=instance)
+    if form.is_valid():
+        instance = form.save(commit=False)
+        instance.save()
+        messages.success(request, "<a href='#'>Item</a> Saved", extra_tags='html_safe')
+        return HttpResponseRedirect(instance.get_absolute_url())
+    context = {
+    "title": instance.title,
+    "instance": instance,
+    "form":form,
+    }
+    return render(request, "post_form.html", context)
 
 
 
 def post_delete(request, slug=None):
+    # Permissions user
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
 	instance = get_object_or_404(Post, slug=slug)
 	instance.delete()
 	messages.success(request, "Successfully deleted")
